moving twitter mentions monitor thread code into it's own scope
